move plotting from helpers into separate file
